// FrontendHackADisc/src/app/admin/interfaces/ResponseCompanyStats.ts
export interface ResponseCompanyStats {
  main_company_id:        number;
  adaptability_to_change_before: number;
  adaptability_to_change_after: number;
  adaptability_to_change_difference: number;
  safe_conduct_before:    number;
  safe_conduct_after:     number;
  safe_conduct_difference: number;
  dynamsim_energy_before: number;
  dynamsim_energy_after:  number;
  dynamsim_energy_difference: number;
  personal_effectiveness_before: number;
  personal_effectiveness_after: number;
  personal_effectiveness_difference: number;
  initiative_before:       number;
  initiative_after:        number;
  initiative_difference:  number;
  working_under_pressure_before: number;
  working_under_pressure_after: number;
  working_under_pressure_difference: number;

}